pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bleach==6.0.0",
    "boto3==1.28.36",
    "django-ratelimit==4.1.0",
    "django-storages==1.14.2",
    "django==5.0.0",
    "djangorestframework==3.14.0",
    "drf-yasg==1.21.7",
    "gunicorn==21.2.0",
    "pillow==10.1.0",
    "pypandoc==1.11",
    "python-docx==0.8.11",
    "requests==2.31.0",
    "djangorestframework-simplejwt==5.3.0",
    "psycopg2-binary==2.9.9",
    "reportlab==4.0.4",
    "openpyxl==3.1.2",
    "python-decouple==3.8",
    "django-cors-headers==4.3.0",
    "django-filter==23.5.0",
    "redis==5.0.1",
    "celery==5.3.4",
]
